requirements.txt
django
django-ckeditor
django-admin-interface
django-ckeditor
django-colorfield
Markdown
pillow
text-unidecode
requests
